+++ src/booleans.h
@@ -28,4 +28,12 @@ pmp_boolean_intersection(
     Eigen::Ref<const compas::RowMatrixXi> &FB);
 
 
+std::tuple<compas::RowMatrixXd, compas::RowMatrixXi>
+pmp_clip(
+    Eigen::Ref<const compas::RowMatrixXd> &VA,
+    Eigen::Ref<const compas::RowMatrixXi> &FA,
+    Eigen::Ref<const compas::RowMatrixXd> &VB,
+    Eigen::Ref<const compas::RowMatrixXi> &FB);
+
+
 #endif /* COMPAS_BOOLEANS_H */